+++ tools/libexec/perl/TTP/Service.pm
@@ -65,9 +65,48 @@ my $Const = {
 ### Public methods
 
 # ------------------------------------------------------------------------------------------------
+# Getter
+# (I):
+# - the list of keys to be considered as an array ref
+# (O):
+# - returns an array ref to the configured commands as provided by TTP::commandByOS(), which may be empty or undef
+
+sub commands {
+	my ( $self, $keys ) = @_;
+
+	my $commands = [];
+
+	# search in 'services'
+	my $serviceKeys = [ 'services', $self->name(), @{$keys} ];
+	push( @{$commands}, TTP::commandByOS( $serviceKeys, { withCommands => true, jsonable => $ep->node() }));
+
+	# search in (deprecated in v4.10) 'Services'
+	$serviceKeys = [ 'Services', $self->name(), @{$keys} ];
+	push( @{$commands}, TTP::commandByOS( $serviceKeys, { withCommands => true, jsonable => $ep->node() }));
+
+	return $commands;
+}
+
+# ------------------------------------------------------------------------------------------------
+# Getter
+# (I):
+# - none
+# (O):
+# - returns true|false, whether the service is enabled, defaulting to true
+
+sub enabled {
+	my ( $self ) = @_;
+
+	my $enabled = $self->var([ 'enabled' ]);
+	$enabled = true if !defined $enabled;
+
+	return $enabled;
+}
+
+# ------------------------------------------------------------------------------------------------
 # Override the 'IJSONable::evaluate()' method to manage the macros substitutions
 # (I):
-# -none
+# - none
 # (O):
 # - this same object
 
@@ -87,7 +126,7 @@ sub evaluate {
 # ------------------------------------------------------------------------------------------------
 # Getter
 # (I):
-# -none
+# - none
 # (O):
 # - returns true|false, whether the service is hidden, defaulting to false
 
@@ -103,7 +142,7 @@ sub hidden {
 # ------------------------------------------------------------------------------------------------
 # Returns the name of the service
 # (I):
-# -none
+# - none
 # (O):
 # - returns the name of the service, or undef
 
@@ -174,15 +213,16 @@ sub var {
 # -------------------------------------------------------------------------------------------------
 # Enumerate the services defined on the node
 # - in ascii-sorted order [0-9A-Za-z]
-# - considering the 'hidden' option
+# - considering and honoring the 'hidden' and 'enabled' options
 # - and call the provided sub for each found
 # (I):
-# - an arguments hash with following keys:
+# - an optional arguments hash with following keys:
 #   > node, the node name or the TTP::Node instance on which the enumeration must be done, defaulting to current execution node
 #   > cb, a code reference to be called on each enumerated service with:
 #     - the TTP::Service instance (remind: it may have not been jsonLoaded)
 #     - this same arguments object
 #   > hidden, whether to also return hidden services, defaulting to false
+#   > disabled: whether to also return the disabled services, defaulting to false
 # (O):
 # - returns a count of enumerated services
 
@@ -192,6 +232,8 @@ sub enum {
 	my $count = 0;
 	my $withHiddens = false;
 	$withHiddens = $args->{hidden} if defined $args->{hidden};
+	my $withDisabled = false;
+	$withDisabled = $args->{disabled} if defined $args->{disabled};
 	my $node = $ep->node();
 	if( defined( $args->{node} )){
 		my $ref = ref( $args->{node} );
@@ -206,20 +248,39 @@ sub enum {
 			$node = $ep->node() if !$node;
 		}
 	}
-	my $cb = $args->{cb};
-	if( $cb && ref( $cb ) eq 'CODE' ){
-		# these are the services defined on this node
-		my $services = $node->var([ 'Services' ]);
-		my @list = sort keys %{$services};
-		foreach my $it ( @list ){
-			my $service = TTP::Service->new( $ep, { service => $it });
-			if( $service && !$service->hidden() || $withHiddens ){
-				$cb->( $service, $args );
-				$count += 1;
+	if( !TTP::errs()){
+		my $cb = $args->{cb};
+		if( $cb && ref( $cb ) eq 'CODE' ){
+			# these are the services defined on this node
+			# via the 'services' property
+			my $services = $node->var([ 'services' ]) || {};
+			my @keys = keys %{$services};
+			foreach my $it ( @keys ){
+				my $service = TTP::Service->new( $ep, { service => $it });
+				if( $service && ( !$service->hidden() || $withHiddens ) && ( $service->enabled() || $withDisabled )){
+					$cb->( $service, $args );
+					$count += 1;
+				}
 			}
+			# via the (deprecated in v4.10) 'Services' property
+			$services = $node->var([ 'Services' ]) || {};
+			@keys = keys %{$services};
+			my $deprecated = 0;
+			foreach my $it ( @keys ){
+				my $service = TTP::Service->new( $ep, { service => $it });
+				if( $service && ( !$service->hidden() || $withHiddens ) && ( $service->enabled() || $withDisabled )){
+					$cb->( $service, $args );
+					$count += 1;
+					$deprecated += 1;
+				}
+			}
+			if( $deprecated && !$node->{_servicesWarned} ){
+				msgWarn( "'Services' property is deprecated in favor of 'services'. You should update your configurations." );
+				$node->{_servicesWarned} = true;
+			}
+		} else {
+			msgErr( __PACKAGE__."::enumerate() expects a 'cb' callback code ref, not found" );
 		}
-	} else {
-		msgErr( __PACKAGE__."::enumerate() expects a 'cb' callback code ref, not found" );
 	}
 	return $count;
 }
@@ -248,6 +309,33 @@ sub finder {
 }
 
 # -------------------------------------------------------------------------------------------------
+# Returns the ordered list of services defined on the specified node
+# (I):
+# - an optional arguments hash with following keys:
+#   > hidden: whether to also return the hidden services, defaulting to false
+#   > disabled: whether to also return the disabled services, defaulting to false
+#   > node: the node to be search the services in, either as a string (the node name) or a TTP::Node ref
+# (O):
+# - returns the list of (maybe hidden) services for the node
+
+sub list {
+	my ( $class, $args ) = @_;
+	$args //= {};
+
+	my $list = [];
+	$args->{cb} = \&_list_cb;
+	$args->{result} = $list;
+	$class->enum( $args ); # ignore returned count
+
+	return $list;
+}
+
+sub _list_cb {
+	my ( $service, $args ) = @_;
+	push( @{$args->{result}}, $service->name());
+}
+
+# -------------------------------------------------------------------------------------------------
 # Constructor
 # (I):
 # - the TTP EP entry point ref
@@ -268,7 +356,7 @@ sub new {
 		# keep the service name
 		$self->{_name} = $args->{service};
 
-		# allowed servicesDirs are configured at site-level
+		# allowed services.confDirs are configured at site-level
 		my $finder = $class->finder();
 		my $findable = {
 			dirs => [ $finder->{dirs}, $args->{service}.$finder->{suffix} ],